refactor: rename page components and update import paths for consistency
